Use loopback address as default for --hosts
In cloud environments, there appear to be periodical outages to hostname
resolution which sometimes causes `localhost` to not resolve to
anything. By using an exact IP address, we remove the need for a hostname
lookup when connecting with the default value for --hosts. This should
improve MaxCtrl and test stability in these environments.